Update 2-1-running-semigrand-mc.ipynb
The deletion reflects that the new chemical potential for the Vacancy specie (or other species) are updated. Previously, a new entry under the key "Vacancy" was being appended to the dictionary. 

I also believe _mu_table needs to be updated by the user or else mu_table still only contains the initial chemical_potentials. The addition reflects this.